// packages/backend/src/usagers/controllers/import/schema-generated-model/UsagersImportUsager.ts
/**
 * This file was automatically generated by joi-to-typescript
 * Do not modify this file manually
 */

import { UsagersImportCivilite } from './attributes';

export interface UsagersImportUsager {
  data?: {
    /**
     * civilite of usager
     */
    civilite: UsagersImportCivilite;
    dateNaissance: string;
    nom: string;
    prenom: string;
  };
  params?: {
    today: Date;
  };
}